requirements.txt
python-dotenv
langchain_groq
langchain_community
langchain-deepseek
langgraph
google-search-results
pydantic
fastapi
uvicorn
streamlit